Complete frontend with backend

+++ Distributed-Rate-limiter-using-token-bucket-algorithm/backend/src/main/java/com/risah/ratelimiter/ratelimiterapp/model/TokenBucket.java
@@ -25,7 +25,7 @@ public class TokenBucket {
         }
         return false;
     }
-    public int getToken(){
+    public long getToken(){
         return tokens;
     }
